\name{friedman.test.with.post.hoc}
\alias{friedman.test.with.post.hoc}
\title{http://www.r-statistics.com/2010/02/post-hoc-analysis-for-friedmans-test-r-code/}
\usage{
  friedman.test.with.post.hoc(formu, data,
    to.print.friedman = T, to.post.hoc.if.signif = T,
    to.plot.parallel = T, to.plot.boxplot = T,
    signif.P = 0.05, color.blocks.in.cor.plot = T,
    jitter.Y.in.cor.plot = F)
}
\description{
  http://www.r-statistics.com/2010/02/post-hoc-analysis-for-friedmans-test-r-code/
}
\author{
  Tal Galili
}

