% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriMatch-package.R
\docType{data}
\name{nmes}
\alias{nmes}
\title{Results from the 1987 National Medical Expenditure Study}
\format{a data frame with 9,708 observations of 12 variables.}
\source{
http://imai.princeton.edu/research/pscore.html
}
\description{
This file was originally prepared by Anders Corr (corr@fas.harvard.edu) who
reports on December 8, 2007 that the resulting numbers closely
match with those reported in the published article. It was later modified by
Jason Bryer (jason@bryer.org) to an R data object to be included in this 
package. See \url{http://imai.princeton.edu/research/pscore.html} for more
information
}
\author{
United States Department of Health and Human Services. 
        Agency for Health Care Policy and Research
}
\references{
National Center For Health Services Research, 1987. National 
        Medical Expenditure Survey. Methods II. Questionnaires and data 
        collection methods for the household survey and the Survey of 
        American Indians and Alaska Natives. National Center for Health 
        Services Research and Health Technology Assessment.
        
        Imai, K., & van Dyk, D.A. (2004). Causal Inference With General 
        Treatment Regimes: Generalizing the Propensity Score, Journal of the 
        American Statistical Association, 99(467), pp. 854-866.
        
        Elizabeth Johnson, E., Dominici, F., Griswold, M., & Zeger, S.L. (2003).
        Disease cases and their medical costs attributable to smoking: An 
        analysis of the national medical expenditure survey. Journal of 
        Econometrics, 112.
}
\keyword{datasets}

