% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.triangle.matches.R
\name{print.trimatch.summary}
\alias{print.trimatch.summary}
\title{Prints the results of [summary.triangle.matches()].}
\usage{
\method{print}{trimatch.summary}(x, ...)
}
\arguments{
\item{x}{results of [summary.triangle.matches()].}

\item{...}{multiple results of [summary.triangle.matches()]. These
must be named. For example, \code{"Method 1" = summary(tmath, outcome)}.}
}
\description{
This is an S3 generic function to print the results of 
[summary.triangle.matches()].
}
