% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utils.R
\name{segment1}
\alias{segment1}
\title{Internal method for plotting.
Position along the left side segement}
\usage{
segment1(d)
}
\arguments{
\item{d}{the distance}
}
\description{
Internal method for plotting.
Position along the left side segement
}

