% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utils.R
\name{segment2}
\alias{segment2}
\title{Internal method for plotting.
Position along the right side segement}
\usage{
segment2(d)
}
\arguments{
\item{d}{the distance}
}
\description{
Internal method for plotting.
Position along the right side segement
}

