\docType{data}
\name{students}
\alias{students}
\title{Results from a new student outreach for increasing credit activity within the
first year of enrollment.}
\format{a data frame with 320 ovservations of 20 variables.}
\description{
  Results from a new student outreach for increasing credit
  activity within the first year of enrollment.
}
\keyword{datasets}

