% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.triangle.psa.R
\name{summary.triangle.psa}
\alias{summary.triangle.psa}
\title{Prints the summary results of the logistic regression models.}
\usage{
\method{summary}{triangle.psa}(object, ...)
}
\arguments{
\item{object}{the results of [trips()].}

\item{...}{currently unused.}
}
\description{
The [trips()] function estimates three separate logistic regression
models for each pair of groups. This function will print a combined table
of the three summaries.
}
