\name{summary.unmatched}
\alias{summary.unmatched}
\title{Provides a summary of unmatched subjects.}
\usage{
  \method{summary}{unmatched} (object, digits = 3, ...)
}
\arguments{
  \item{object}{results of \code{\link{unmatched}}}

  \item{digits}{number of digits to print.}

  \item{...}{currently unused.}
}
\value{
  a list of summary results.
}
\description{
  Will return as a list and print the percentage of total
  unmatched rows and percent by treatment.
}

