\docType{data}
\name{tutoring}
\alias{tutoring}
\title{Results from a study examining the effects of tutoring services on course grades.}
\format{a data frame with 1,381 ovservations of 17 variables.}
\description{
  Results from a study examining the effects of tutoring
  services on course grades.
}
\keyword{datasets}

