\name{merge.triangle.matches}
\alias{merge.triangle.matches}
\title{Merges outcomes with the matched set.}
\usage{
  merge.triangle.matches(x, y, ...)
}
\arguments{
  \item{x}{the result of \code{\link{trimatch}}}

  \item{y}{another data frame or vetor to merge with.}

  \item{...}{unused}
}
\value{
  \code{x} with the additional column(s) added.
}
\description{
  Merges outcomes with the matched set.
}

