% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multibalance.R
\name{multibalance.plot}
\alias{multibalance.plot}
\title{Multiple covariate balance assessment plot.}
\usage{
multibalance.plot(tpsa, tmatch, grid = TRUE, cols)
}
\arguments{
\item{tpsa}{results of \code{\link{trips}}.}

\item{tmatch}{results of \code{\link{trimatch}}.}

\item{grid}{if TRUE, then a grid of three plots for each model will be displayed.}

\item{cols}{character vector of covariates (i.e. column names) from the original 
data to include in the plot. By default all covariates used in the
logistic regression model are used.}
}
\value{
a \code{ggplot2} figure.
}
\description{
A graphic based upon \code{\link{cv.bal.psa}} function in the \code{PSAgraphics}
package. This graphic plots the effect sizes for multiple covariates before and
after propensity score adjustment.
}
