% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.parallel.R
\name{parallel.plot}
\alias{parallel.plot}
\title{Parallel coordinate plot for the three groups and dependent variable.}
\usage{
parallel.plot(tmatch, outcome)
}
\arguments{
\item{tmatch}{results from \code{\link{trimatch}}.}

\item{outcome}{vector of the outcome}
}
\description{
Creates a \code{ggplot2} figure of a parallel coordinate plot.
}

