\name{plot.multibalance}
\alias{plot.multibalance}
\title{Multiple covariate blance assessment plot.}
\usage{
  plot.multibalance(tpsa, covs, grid = FALSE)
}
\description{
  A graphic based upon \code{\link{cv.bal.psa}} function in
  the \code{PSAgraphics} package. This graphic plots the
  effect sizes for multiple covariated before and after
  propensity score andjustement.
}

