\name{plot.triangle.psa}
\alias{plot.triangle.psa}
\title{Triangle plot showing the fitted values (propensity scores) for three different
models.}
\usage{
  plot.triangle.psa(tpsa, point.alpha = 0.3,
    point.size = 5, legend.title = "Treatment",
    text.size = 4, draw.edges = FALSE,
    draw.segments = TRUE, edge.alpha = 0.2,
    edge.colour = "grey",
    edge.labels = c("Model 1", "Model 2", "Model 3"), ...)
}
\arguments{
  \item{tpsa}{the results from}
}
\description{
  Triangle plot showing the fitted values (propensity
  scores) for three different models.
}
\seealso{
  triangle.psa
}

