% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.balance.R
\name{print.balance.plots}
\alias{print.balance.plots}
\title{Print the results of \code{\link{balance.plot}} for a data frame of covariates.}
\usage{
\method{print}{balance.plots}(x, ...)
}
\arguments{
\item{x}{the results of \code{\link{balance.plot}} when a data frame is specified.}

\item{...}{parameters passed to \code{\link{plot.balance.plots}} and
\code{\link{summary.balance.plots}}.}
}
\description{
Print the results of \code{\link{balance.plot}} for a data frame of covariates.
}

