\name{print.triangle.plot}
\alias{print.triangle.plot}
\title{Print method for \code{\link{plot.triangle.psa}}. The primary purpose is to
suppress the "Removed n rows containing missing values" warning printed
by \code{ggplot2}.}
\usage{
  print.triangle.plot(p, ...)
}
\arguments{
  \item{p}{a plot from \code{\link{plot.triangle.psa}}.}

  \item{...}{other parameters passed to ggplot2.}
}
\description{
  Print method for \code{\link{plot.triangle.psa}}. The
  primary purpose is to suppress the "Removed n rows
  containing missing values" warning printed by
  \code{ggplot2}.
}

