\name{print.trimatch.summary}
\alias{print.trimatch.summary}
\title{Prints the results of \code{\link{summary.triangle.matches}}.}
\usage{
  \method{print}{trimatch.summary} (x, ...)
}
\arguments{
  \item{x}{results of
  \code{\link{summary.triangle.matches}}.}

  \item{...}{currenlty unused.}
}
\description{
  This is an S3 generic function to print the results of
  \code{\link{summary.triangle.matches}}.
}

