\name{summary.triangle.psa}
\alias{summary.triangle.psa}
\title{Prints the summary results of the logistic regression models.}
\usage{
  summary.triangle.psa(object, ...)
}
\arguments{
  \item{object}{the results of \code{\link{trips}}.}

  \item{...}{currently unused.}
}
\description{
  The \code{\link{trips}} function estimates three separate
  logistic regression models for each pair of groups. This
  function will print a combined table of the three
  summaries.
}

