% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmatched.R
\name{unmatched}
\alias{unmatched}
\title{Returns rows from [trips()] that were not matched by [trimatch()].}
\usage{
unmatched(tmatch)
}
\arguments{
\item{tmatch}{the results of [trimatch()].}
}
\value{
a data frame of unmatched rows.
}
\description{
This function returns a subset of [trips()] that were not matched by
[trimatch()]. All data frame methods work with the returned object
but special `summary` function will provided relevant information.
}
